[src] Build and ship a very simple implementation-only version of OpenTK-1.0.dll for Mac Catalyst.
This version of OpenTK-1.0.dll only contains type forwarders for
System.Drawing.Color and System.Drawing.KnownColor. Eventually it may also
contain stubs that throw PlatformNotSupportedExceptions.

The main problem we're trying to solve is that we want to make the version of
Xamarin.Essentials built for Xamarin.iOS work in Mac Catalyst, and it
references OpenTK-1.0.dll, which means that we'll show an error at build time
if we can't find OpenTK-1.0.dll. This way there will be an OpenTK-1.0.dll, but
it's only available as an implementation assembly, which means that it's not
possible to reference it directly from a Mac Catalyst project.