[CI][VSTS] Use the creds store in git for https. (#10535)
VSTS is using https and not ssh to clone the repos, this means than if
we have any bash that uses git, it will fail. Cred store fixes this by
removing the need to pass the token in all bash scripts.

We remove allways the cred store once we are done.