[src] Move enums into files that are processed by the generator
The generator has been able to read and re-generate enums for a while.
However not all enums were in file processed by the generator. That was
not a huge problem since the output was identical, unless this was a
"smart enum".

Supporting .net5+ `[Uns|S]upportedOSPlatformAttribute` on API requires
a lot of changes. Generated bindings are much easier since the generator
will output the new attributes for dotnet and keep the older ones for
_legacy_.

So moving the enums to generated code means a lot less code to update
for dotnet - and means additional enums (new Xcode) won't have to include
both new and legacy attributes.

The number of manual updates (for iOS) is about 1500 less (5792 down to
4219) than before.

Helps https://github.com/spouliot/xamarin-macios/pull/new/gh10170