[tools] Fix issue with default verbosity. (#10643)
* Store verbosity in only one place (ErrorHelper.Verbosity).
* Initialize to the default value in the Driver's static constructor.

There was an issue where the default verbosity wasn't set everywhere, because
ErrorHelper.Verbosity was only set when Driver.Verbosity was assigned, which
wouldn't happen if no -v/-q was passed on the command line. The result would
be that any files in $HOME that specifies a default verbosity would not be
taken into account in ErrorHelper.