[maccatalyst] Increase iOS compatibility by removing `XAMCORE_3_0`
`XAMCORE_3_0` was added in tvOS and watchOS where binary compatibility
was not a consideration. The work to update source code for `XAMCORE_3_0`
was also minimal.

The former is not the always the case for MacCatalyst, while the latter
is still true.

Fix https://github.com/xamarin/xamarin-macios/issues/11237 and https://github.com/xamarin/xamarin-macios/issues/11236