[CI] Remove Mac OS devices that will not happen. (#11276)
Remove those devices of Mac OS X that will not happen because we cannot
run the agent in such and old device. This is due to the fact that the
agent runs using c# and we do not support such and old OS.

Also add a new stage to run tests on catalina. ATM older OS as not
working, but Catalina should work with our trusted pool.

Co-authored-by: Manuel de la Pena <mandel@microsoft.com>