[Actions] Clean data after a PR is closed.
To keep the github  webpage small we remove the data once the PR has
been closed (that means closed or merged). The data will be kept in
vsdrops, the only links that will stop working are those from the github
static webpage.

Since we are interested in making the reviewer life better this is a
good compromise.

How to avoif merge conflicts:

We want to make sure we avoid merge conflicts, this is done using two
locks:

1. macios.ci-cleanup - Lock in the xamarin-macios action that will
   ensure that we create a new branch to cleanup the data.
2. ci-results - Lock in the macios.ci action that will merge branhces
   one by one.

This repo will create a new pr that cleans the data, the macios.ci repo
will execute an action when this new branch is created and will merge it
as soon as possible making sure that previos created branches are
merged.