[Actions] Remove broken action. (#11819)
Actions that are triggered from forks do not have access to the secrets
making the action fail most of the time. We have moved to a github
action in the static page repo that will clean results of closed PRs
everyday. We remove this action because is broken and not needed.