[dotnet] Add support for generating a binary version of runtimeconfig.json. Fixes #11745.
Use Mono's RuntimeConfigParserTask to parse the *.runtimeconfig.json file and
produce a binary version of it. This way Mono doesn't have to do json parsing
at runtime, which makes it possible to make the runtimer smaller and execution
faster.

This also means implementing support for finding the on-disk location of the
file at runtime, and passing it to mono.

Ref: https://github.com/dotnet/runtime/blob/01b7e73cd378145264a7cb7a09365b41ed42b240/docs/design/mono/mobile-runtimeconfig-json.md

Fixes https://github.com/xamarin/xamarin-macios/issues/11745.